+++ hpc/nvidia/cuda/cuda_compat.h
@@ -194,10 +194,11 @@
 # endif
 # ifdef HAVE_CUPTI
 #  include <cupti.h>
-#  include <cupti_target.h>
 #  include <cupti_pmsampling.h>
 #  include <cupti_profiler_target.h>
 #  include <cupti_profiler_host.h>
+#  include <cupti_range_profiler.h>
+#  include <cupti_target.h>
 # endif
 # define CUNAME	"NVIDIA CUDA"
 #endif